refactor amp/bento-timeago
