+++ extensions/amp-timeago/1.0/amp-timeago.js
@@ -1,13 +1,21 @@
 import {isExperimentOn} from '#experiments';
 
+import {AmpPreactBaseElement} from '#preact/amp-base-element';
+
 import {userAssert} from '#utils/log';
 
-import {BaseElement} from './base-element';
+import {
+  Component,
+  layoutSizeDefined,
+  props,
+  updatePropsForRendering,
+  usesShadowDom,
+} from './element';
 
 /** @const {string} */
 const TAG = 'amp-timeago';
 
-class AmpTimeago extends BaseElement {
+class AmpTimeago extends AmpPreactBaseElement {
   /** @override */
   isLayoutSupported(layout) {
     userAssert(
@@ -17,7 +25,24 @@ class AmpTimeago extends BaseElement {
     );
     return super.isLayoutSupported(layout);
   }
+
+  /** @override */
+  updatePropsForRendering(props) {
+    updatePropsForRendering(props);
+  }
 }
+
+/** @override */
+AmpTimeago['Component'] = Component;
+
+/** @override */
+AmpTimeago['layoutSizeDefined'] = layoutSizeDefined;
+
+/** @override */
+AmpTimeago['props'] = props;
+
+/** @override */
+AmpTimeago['usesShadowDom'] = usesShadowDom;
 
 AMP.extension(TAG, '1.0', (AMP) => {
   AMP.registerElement(TAG, AmpTimeago);
