+++ extensions/amp-timeago/1.0/base-element.js
@@ -1,46 +1,28 @@
-import {parseDateAttrs as parseDateAttrsBase} from '#core/dom/parse-date-attributes';
-
 import {PreactBaseElement} from '#preact/base-element';
 
-import {BentoTimeago} from './component';
+import {
+  Component,
+  layoutSizeDefined,
+  props,
+  updatePropsForRendering,
+  usesShadowDom,
+} from './element';
 
 export class BaseElement extends PreactBaseElement {
   /** @override */
   updatePropsForRendering(props) {
-    props['placeholder'] = props['children'];
+    updatePropsForRendering(props);
   }
 }
 
 /** @override */
-BaseElement['Component'] = BentoTimeago;
+BaseElement['Component'] = Component;
 
 /** @override */
-BaseElement['layoutSizeDefined'] = true;
+BaseElement['layoutSizeDefined'] = layoutSizeDefined;
 
 /** @override */
-BaseElement['props'] = {
-  'children': {passthroughNonEmpty: true},
-  'cutoff': {attr: 'cutoff', type: 'number'},
-  'datetime': {
-    attrs: ['datetime', 'timestamp-ms', 'timestamp-seconds', 'offset-seconds'],
-    parseAttrs: parseDateAttrs,
-  },
-  'locale': {attr: 'locale'},
-};
+BaseElement['props'] = props;
 
 /** @override */
-BaseElement['usesShadowDom'] = true;
-
-/**
- * @param {!Element} element
- * @return {?number}
- * @throws {UserError} when attribute values are missing or invalid.
- * @visibleForTesting
- */
-export function parseDateAttrs(element) {
-  return parseDateAttrsBase(element, [
-    'datetime',
-    'timestamp-ms',
-    'timestamp-seconds',
-  ]);
-}
+BaseElement['usesShadowDom'] = usesShadowDom;
